+++ toolchain/scripts/Builder-Manager.ps1
@@ -57,6 +57,7 @@ function Main {
   Write-Host "--------------------------------------------"
   Write-Host ""
 
+  Import-Module "$($config.project.neocorePath)\toolchain\scripts\modules\models\project-name.ps1"
   Import-Module "$($config.project.neocorePath)\toolchain\scripts\modules\models\rule.ps1"
   Import-Module "$($config.project.neocorePath)\toolchain\scripts\modules\utils\logger.ps1"
   Import-Module "$($config.project.neocorePath)\toolchain\scripts\modules\services\installers\sdk.ps1"
@@ -65,6 +66,7 @@ function Main {
   Import-Module "$($config.project.neocorePath)\toolchain\scripts\modules\utils\remove-project.ps1"
 
   Model-Rule -Rule $($buildConfig.rule)
+  Model-Project-Name -Name $($Config.project.name)
 
   $raineProcessName = [System.IO.Path]::GetFileNameWithoutExtension($Config.project.emulator.raine.exeFile)
   $mameProcessName = [System.IO.Path]::GetFileNameWithoutExtension($Config.project.emulator.mame.exeFile)